angularWorkshop/Testing/src/app/components/testinputoutput/testinputoutput.component.spec.ts
import { DebugElement } from '@angular/core';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { By } from '@angular/platform-browser';

import { TestinputoutputComponent, User } from './testinputoutput.component';

describe('TestinputoutputComponent', () => {
  let component: TestinputoutputComponent;
  let fixture: ComponentFixture<TestinputoutputComponent>;

  // DebugElement is an Angular class that contains all kinds of references and
  //  methods relevant to investigate an element as well as component

  //  you can pass id as By.css('#shan') , class as By.css('.shan') , 
  // or you can also elements with ways such as By.css('div') or By.css('some-app-component')
  
  let submitEl: DebugElement;
  let loginEl: DebugElement;
  let passwordEl: DebugElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [TestinputoutputComponent]
    })
      .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(TestinputoutputComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
    submitEl = fixture.debugElement.query(By.css('button'));
    loginEl = fixture.debugElement.query(By.css('input[type=email]'));
    passwordEl = fixture.debugElement.query(By.css('input[type=password]'));
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  it('Setting enabled to true enables the submit button', () => {
    component.enabled = true;
    fixture.detectChanges();
    expect(submitEl.nativeElement.disabled).toBeFalsy();
});

it('Entering email and password emits loggedIn event', () => {
    let user: User;
    loginEl.nativeElement.value = "test@example.com";
    passwordEl.nativeElement.value = "123456";

    // Subscribe to the Observable and store the user in a local variable.
    component.loggedIn.subscribe((value) => user = value);

    // This sync emits the event and the subscribe callback gets executed above

    // there is not handlers for click event 
                                       // that angular added to this element 
                                       // so test method won't be called
    submitEl.triggerEventHandler('click', null);

    // Now we can check to make sure the emitted value is correct
    expect(user.email).toBe("test@example.com");
    expect(user.password).toBe("123456");
});
});
